Twitter/requirements.txt

Flask-SQLAlchemy==2.0

Flask-WTF==0.10.3

Jinja2 
MarkupSafe==0.23

PyMySQL==0.6.3


WTForms==2.0.1

Werkzeug==0.14.1
argparse==1.2.1


itsdangerous==0.24


boto==2.28.0

nltk==3.3

py2neo==4.0.0

textblob==0.15.1

vaderSentiment==3.2.1

Flask==1.0.2

neo4j-driver==1.6.0

tweepy==3.6.0

SQLAlchemy==1.2.9
DateTime==4.2